fix: Implement delete correctly by incorporating sentinel nodes
Programming is a tough job. Sometimes the greatest victories boil down
to writing a commit message in a project that no one will ever use in a message no one will ever see. But this commit is a victory none-the-less. Sometimes you write a lot of code, just to figure out it won't work, and the have to run `git stash`. Theres no record of all the work
you did, all the toil you had. And that makes it feel like less. But
that work was still done, and this message still written,